report: add Util.prepareReportResult method (#5766)
