+++ lighthouse-core/test/report/html/renderer/report-renderer-test.js
@@ -21,7 +21,7 @@ let PerformanceCategoryRenderer = null;
 const CriticalRequestChainRenderer = require(
     '../../../../report/html/renderer/crc-details-renderer.js');
 const ReportRenderer = require('../../../../report/html/renderer/report-renderer.js');
-const sampleResults = require('../../../results/sample_v2.json');
+const sampleResultsOrig = require('../../../results/sample_v2.json');
 
 const TIMESTAMP_REGEX = /\d+, \d{4}.*\d+:\d+/;
 const TEMPLATE_FILE = fs.readFileSync(__dirname +
@@ -29,6 +29,7 @@ const TEMPLATE_FILE = fs.readFileSync(__dirname +
 
 describe('ReportRenderer', () => {
   let renderer;
+  let sampleResults;
 
   beforeAll(() => {
     global.URL = URL;
@@ -57,7 +58,7 @@ describe('ReportRenderer', () => {
     const detailsRenderer = new DetailsRenderer(dom);
     const categoryRenderer = new CategoryRenderer(dom, detailsRenderer);
     renderer = new ReportRenderer(dom, categoryRenderer);
-    sampleResults.reportCategories = Object.values(sampleResults.categories);
+    sampleResults = Util.prepareReportResult(sampleResultsOrig);
   });
 
   afterAll(() => {
